Solution for interface segregation.
