--- src/main/java/RestaurantInterface.java
@@ -1,8 +1,0 @@
-public interface RestaurantInterface {
-
-  void acceptOnlineOrder();
-  void acceptTelephoneOrder();
-  void acceptWalkInCustomerOrder();
-  void payOnline();
-  void payInPerson();
-}
