+++ src/main/java/org/example/epam/jat/volha/oop/AcceptOnlineOrderInterface.java
@@ -1,0 +1,5 @@
+package org.example.epam.jat.volha.oop;
+
+public interface AcceptOnlineOrderInterface extends AcceptOrderInterface {
+  void acceptOnlineOrder();
+}
