+++ src/main/java/org/example/epam/jat/volha/oop/AcceptTelephoneOrderInterface.java
@@ -1,0 +1,5 @@
+package org.example.epam.jat.volha.oop;
+
+public interface AcceptTelephoneOrderInterface extends AcceptOrderInterface {
+  void acceptTelephoneOrder();
+}
