+++ src/main/java/org/example/epam/jat/volha/oop/AcceptWalkInCustomerOrderInterface.java
@@ -1,0 +1,5 @@
+package org.example.epam.jat.volha.oop;
+
+public interface AcceptWalkInCustomerOrderInterface extends AcceptOrderInterface {
+  void acceptWalkInCustomerOrder();
+}
