+++ src/main/java/org/example/epam/jat/volha/oop/PayInPersonInterface.java
@@ -1,0 +1,5 @@
+package org.example.epam.jat.volha.oop;
+
+public interface PayInPersonInterface {
+  void payInPerson();
+}
