+++ src/main/java/org/example/epam/jat/volha/oop/PayInterface.java
@@ -1,0 +1,4 @@
+package org.example.epam.jat.volha.oop;
+
+public interface PayInterface extends RestaurantInterface{
+}
