+++ src/main/java/org/example/epam/jat/volha/oop/PayOnlineInterface.java
@@ -1,0 +1,5 @@
+package org.example.epam.jat.volha.oop;
+
+public interface PayOnlineInterface {
+  void payOnline();
+}
