+++ src/main/java/org/example/epam/jat/volha/oop/RestaurantInterface.java
@@ -1,0 +1,6 @@
+package org.example.epam.jat.volha.oop;
+
+//I wish I could specify that e.g.   void acceptTelephoneOrder(); is some implementation of acceptOrder method of
+// AcceptOrder interface, however, I've not thought out how to do that.
+public interface RestaurantInterface {
+}
